% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{extracted_bp}
\alias{extracted_bp}
\title{Extracted baseline parameters for the vignette that reproduces the publication results.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 3 columns.}
\usage{
data(extracted_bp)
}
\description{
This contains the mean and standard deviation of (a) the standing systolic blood
pressures taken during orthostatic vital signs, and (b) the CAPS-IV PTSD assessment
total severity score.
}
\keyword{datasets}
