% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{extracted_rp}
\alias{extracted_rp}
\title{Extracted response parameters for the vignette that reproduces the publication results.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 5 columns.}
\usage{
data(extracted_rp)
}
\description{
This contains a set of maximum, displacement, rate and standard deviation values that, when
fed into the function \link{modgompertz} produces a combination of response factors
(biologic response to the drug, br here or BR in the publication; expectancy related response,
pb here or ER in the publication; and the non-treatment related time-varying factor, tv here
or TR in the publication) that together could plausibly describe our existing RCT data set.
This set of parameters was designed to make no other particular assumptions about how these
factors would perform - to be a 'tabula rasa' parameter set - where e.g. the timecourse and
maximum values of the expectancy related and tv/TR factors were assumed to be equal.
}
\keyword{datasets}
