% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generate simulated data}
\usage{
generateData(
  modelparam,
  respparam,
  blparam,
  trialdesign,
  empirical,
  makePositiveDefinite,
  seed = NA
)
}
\arguments{
\item{modelparam}{a datatable with a single entry per named column:
\itemize{
 \item{\code{N}}{  Number of simulated participants}
 \item{\code{c.bm}}{  Correlation between the biomarker and the biologic response}
 \item{\code{carryover_t1half}}{  Halflife of the carryover effect}
 \item{\code{c.tv}}{  Autocorrelation for the tv factor across timepoints}
 \item{\code{c.pb}}{ Autocorrelation for the pb factor across timepoints}
 \item{\code{c.br}}{  Autocorrelation for the br factor across timepoints}
 \item{\code{c.cf1t}}{  Correlation between different factors at a single timepoint}
 \item{\code{c.cfct}}{  Correlation between different factors at different timepoints}
}}

\item{respparam}{Data table with 5 columns and 3 rows. The first row lists the 3 factors
  defining the response trajectories (tv, pb, and br). The rest of the columns are the
  input parameters for the \link{modgompertz} function:
\itemize{
 \item{\code{max}}{  Maximum response value}
 \item{\code{disp}}{  Displacement}
 \item{\code{rate}}{  Rate}
 \item{\code{sd}}{  Standard deviation}
}}

\item{blparam}{Data table with 3 colums and 3 rows. First row lists the varaibles being
defined, which consist of \code{BL} for the baseline value of the outcome measure, and
\code{bm} for the biomarker values. Remaining columns:
\itemize{
  \item{\code{m}}{  Mean}
  \item{\code{sd}}{  Standard deviation}
}}

\item{trialdesign}{The information defining a single path through the clinical trial
design, as defined by the \code{output$trialpaths} output of \link{buildtrialdesign}}

\item{empirical}{Should the correlations be empirical? (Usually \code{FALSE} unless
you're doing something odd for testing purposes)}

\item{makePositiveDefinite}{Should the covariance matrix be forced to be positive definite?
(Usually yes, although you may want to turn this off at times to test the impact of
this step on your covariance sturcture)}

\item{seed}{Randomization seed (defaults to NA)}
}
\value{
A \code{dat} file that contains both the total symptom scores at each timepoint
  and also all the individual factors that were used to generate those total scores
}
\description{
\code{generateData} generates a set of simulated clinical trials' data
}
\details{
This is the core data simulation code. See vignettes for additional details.
This function is primarily called by \link{generateSimulatedResults}.
}
\examples{
  #See vignettes for examples of how to use this
}
