% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotModelingResults}
\alias{PlotModelingResults}
\title{Plot factor trajectories}
\usage{
PlotModelingResults(
  data,
  param2plot,
  param2vary,
  param2hold,
  param2nothold,
  labelcode
)
}
\arguments{
\item{data}{the \code{$results} component of the output of \link{generateSimulatedResults}}

\item{param2plot}{What value to plot as the heatmap out put. Can be power,
mean_frac_NA, mean_beta, mean_betaSE, sd_beta}

\item{param2vary}{4 params to vary and plot across the created space}

\item{param2hold}{Params to keep constant}

\item{param2nothold}{Things to let vary freely (if e.g. they are linked to another
intentionally varied parameter)}

\item{labelcode}{A mapping between interal variable names and figure labels}
}
\value{
a ggplot object containing the plot
}
\description{
\code{PlotModelingResults} plots basic output relating to the peformance of each
trial design across a parameter space
}
\examples{
data(results_core)
  param2vary<-c("trialdesign","N","c.bm","censorparamset")
  param2hold<-data.table(blparamset=1,respparamset=1,carryover_t1half=0)
  param2nothold<-c("c.cfct","modelparamset")
  param2plot<-"power" # Options: power, mean_frac_NA, mean_beta, mean_betaSE, sd_beta...
  p1<-PlotModelingResults(simresults,param2plot,param2vary,param2hold,param2nothold)
 # See vignettes for additional details
}
