% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{data}
\name{results_core}
\alias{results_core}
\title{The results of the vigenette Part 1 used to plot basic paramater space}
\format{An object of class \code{list} of length 2.}
\usage{
data(results_core)
}
\description{
The results of the vigenette Part 1 used to plot basic paramater space
}
\keyword{datasets}
