% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_analysis.R
\name{lme_analysis}
\alias{lme_analysis}
\title{LME Analysis}
\usage{
lme_analysis(trialdesign_set, dat, op)
}
\arguments{
\item{trialdesign_set}{The set of all pathways for the clinical trial}

\item{dat}{A dat file as generated by \link{generateData}}

\item{op}{an options file with the following components:
\itemize{
  \item{\code{op$useDE=TRUE}}{  Binary: should the LME model use expectnacy
    information if available?}
  \item{\code{op$t_random_slope=FALSE}}{  Binary: should the LME model use
    random slopes, as well as random intercepts, for participants?}
}}
}
\value{
Returns a one row data table with components:
  \itemize{
    \item{\code{out$beta}}{  The coefficient corresponding to an interaction
      between biomarker and outcome}
    \item{\code{out$betaSE}}{  Standard error of the above coefficient}
    \item{\code{out$p}}{  p-value conveying the statistical significance of the coefficient}
    \item{\code{out$issingular}}{  Was the model fit flagged as singular?}
    \item{\code{out$warning}}{  Any warnings from \code{fitMsgs}}
  }
}
\description{
\code{lme_analysis} analyzes the data from a single simulated trial
}
\details{
Analyzes the simulated data (including all paths) from a single simulated
trial. Most often called by \code{generateSimualtedResults}. See vignettes
for additional information.
}
\examples{
# See vignettes for examples of how to call from \link{generateSimualtedResults}
}
